pyproject.toml
[project]
name = "avg-portfolio-tracker"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "asyncpg>=0.31.0",
    "hyperliquid-python-sdk>=0.4.0",
    "pandas>=2.0.0",
    "python-dotenv>=1.0.0",
    "requests>=2.31.0",
    "streamlit>=1.30.0",
    "web3>=6.0.0",
]
